requirements.txt
coqui-tts>=0.27.0
pyperclip>=1.11.0
spacy[ja]>=3.8.7
pynput>=1.8.1
pydub>=0.25.1
pyaudio>=0.2.14